refactor(bridge): abstract monitoring logic
